chore: separate cache and contracts opts
