--- src/iam/models.ts
@@ -1,7 +1,0 @@
-export interface ChainConfig {
-  rpcUrl: string,
-  ensRegistryAddress: string,
-  ensResolverAddress: string,
-  didContractAddress: string,
-  cacheServerUrl: string;
-}
